+++ remind_mfa/cement/cement_model.py
@@ -6,7 +6,7 @@ from remind_mfa.cement.cement_definition import get_definition
 from remind_mfa.cement.cement_mfa_system_historic import (
     InflowDrivenHistoricCementMFASystem,
 )
-from remind_mfa.common.scenarios import ParameterExtensionManager
+from remind_mfa.common.parameter_extension import ParameterExtensionManager
 from remind_mfa.cement.cement_mfa_system_historic import InflowDrivenHistoricCementMFASystem
 from remind_mfa.cement.cement_mfa_system_future import StockDrivenCementMFASystem
 from remind_mfa.cement.cement_data_reader import CementDataReader
